# openapi_v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Invest Solo API",
    "description": "Quantitative fundamental analysis API for PEA and Global investment strategies.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/watchlist/": {
      "get": {
        "tags": [
          "watchlist"
        ],
        "summary": "Get Watchlist",
        "description": "Return the full watchlist with live data and three-horizon scoring.\n\nAll three horizon scores (score_lt, score_mt, score_st) are present on\nevery item. Use signal_lt / signal_mt / signal_st as appropriate for the\ninvestment timeframe.",
        "operationId": "get_watchlist_api_watchlist__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "watchlist"
        ],
        "summary": "Add To Watchlist",
        "description": "Add a ticker to the watchlist.",
        "operationId": "add_to_watchlist_api_watchlist__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWatchlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Add To Watchlist Api Watchlist  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/watchlist/{item_id}": {
      "delete": {
        "tags": [
          "watchlist"
        ],
        "summary": "Remove From Watchlist",
        "description": "Remove an item from the watchlist.",
        "operationId": "remove_from_watchlist_api_watchlist__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Remove From Watchlist Api Watchlist  Item Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/screener/run": {
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Run Screen",
        "description": "Run a horizon-aware custom screen.\n\nWhen ``preset`` is set in the request body, it takes precedence over\n``custom_filters``. Otherwise ``custom_filters`` (if provided) are\npassed directly to ``apply_filters``.",
        "operationId": "run_screen_api_screener_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/screener/preset/{preset_name}": {
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Run Preset",
        "description": "Run a named M8 preset screen.\n\n``preset_name`` must be one of the 9 keys in PRESET_REGISTRY\n(e.g. ``LT_QUALITY_COMPOUNDER``, ``ST_MOMENTUM_QUALITY``).",
        "operationId": "run_preset_api_screener_preset__preset_name__post",
        "parameters": [
          {
            "name": "preset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/screener/tickers": {
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Score Ticker List",
        "description": "Fetch and score a custom list of tickers (live data from yfinance).",
        "operationId": "score_ticker_list_api_screener_tickers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickerListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddWatchlistRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "minLength": 1,
            "title": "Ticker",
            "description": "Stock ticker symbol"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Free-text notes"
          }
        },
        "type": "object",
        "required": [
          "ticker"
        ],
        "title": "AddWatchlistRequest",
        "description": "Request body for adding a stock to the watchlist."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PresetRunRequest": {
        "properties": {
          "pea_only": {
            "type": "boolean",
            "title": "Pea Only",
            "description": "Restrict to PEA-eligible stocks only",
            "default": false
          },
          "top_n": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Top N",
            "description": "Max results to return",
            "default": 50
          }
        },
        "type": "object",
        "title": "PresetRunRequest",
        "description": "Request body for running a preset screen."
      },
...
      "ScreenerRequest": {
        "properties": {
          "horizon": {
            "type": "string",
            "enum": [
              "long_term",
              "medium_term",
              "short_term"
            ],
            "title": "Horizon",
            "description": "Scoring horizon that drives sort order and gate filtering",
            "default": "long_term"
          },
          "preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset",
            "description": "Named preset from PRESET_REGISTRY (overrides custom_filters when set)"
          },
          "custom_filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Filters",
            "description": "Ad-hoc filter key-value pairs passed directly to apply_filters"
          },
          "pea_only": {
            "type": "boolean",
            "title": "Pea Only",
            "description": "Restrict to PEA-eligible stocks only",
            "default": false
          },
          "sort_by": {
            "type": "string",
            "title": "Sort By",
            "description": "Column to sort by. 'score' maps to score_<horizon>. Other accepted values: 'pe', 'roe', 'market_cap', 'div_yield', 'revenue_growth', 'altman_z', 'piotroski_f'.",
            "default": "score"
          },
          "sort_desc": {
            "type": "boolean",
            "title": "Sort Desc",
            "description": "Sort descending",
            "default": true
          },
          "limit": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Max results to return",
            "default": 50
          }
        },
        "type": "object",
        "title": "ScreenerRequest",
        "description": "Request body for running a custom horizon-aware screen."
      },
      "ScreenerResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ScreenerResultItem"
            },
            "type": "array",
            "title": "Results"
          },
          "summary": {
            "$ref": "#/components/schemas/ScreenerSummary"
          }
        },
        "type": "object",
        "required": [
          "results",
          "summary"
        ],
        "title": "ScreenerResponse",
        "description": "Full screener API response."
      },
...
      "TickerListRequest": {
        "properties": {
          "tickers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Tickers"
          },
          "horizon": {
            "type": "string",
            "title": "Horizon",
            "description": "Horizon for sort order",
            "default": "long_term"
          }
        },
        "type": "object",
        "required": [
          "tickers"
        ],
        "title": "TickerListRequest",
        "description": "Request body for scoring a custom list of tickers."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WatchlistItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "added_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Added Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "current_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price"
          },
          "pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe"
          },
          "pb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb"
          },
          "roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe"
          },
          "score_lt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Lt",
            "description": "Long-term composite score 0-100"
          },
          "score_mt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Mt",
            "description": "Medium-term composite score 0-100"
          },
          "score_st": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score St",
            "description": "Short-term composite score 0-100"
          },
          "signal_lt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Lt",
            "description": "Long-term investment signal"
          },
          "signal_mt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Mt",
            "description": "Medium-term investment signal"
          },
          "signal_st": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal St",
            "description": "Short-term investment signal"
          },
          "piotroski_f": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Piotroski F"
          },
          "altman_z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altman Z"
          },
          "graham_number": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graham Number"
          },
          "graham_mos": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graham Mos"
          },
          "dcf_mos_mid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dcf Mos Mid"
          },
          "pea_eligible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pea Eligible"
          },
          "pea_pme_eligible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pea Pme Eligible"
          },
          "analyst_rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Rating"
          },
          "analyst_target_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Target Price"
          },
          "forward_pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forward Pe"
          },
          "peg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peg"
          },
          "fifty_two_week_high_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fifty Two Week High Pct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticker"
        ],
        "title": "WatchlistItem",
        "description": "A single enriched watchlist entry with live data and three-horizon scoring."
      },
      "WatchlistResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WatchlistItem"
            },
            "type": "array",
            "title": "Items"
          },
          "last_refreshed": {
            "type": "string",
            "title": "Last Refreshed"
          }
        },
        "type": "object",
        "required": [
          "items",
          "last_refreshed"
        ],
        "title": "WatchlistResponse",
        "description": "Full watchlist API response."
      }
    }
  }
}